alerts/backend/requirements.txt
fastapi==0.116.1
httpx==0.28.1
pydantic==2.11.7
python-dotenv==1.1.0
tinytuya==1.16.1
uvicorn[standard]==0.35.0